policy/requirements.txt
gymnasium
stable-baselines3
torch
pandas
numpy
psycopg2-binary
ccxt
python-dotenv
